# back/openapi.yaml
openapi: 3.0.3
info:
  title: IS Course Work Back API
  description: >-
    REST API that powers the restaurant project. It exposes
    endpoints to manage employees, dishes, ingredients, tables, orders, bills,
    feedback, reports, and employee wallets. Most operations require a JWT that
    encodes the employee id and position (role).
  version: 1.0.0
servers:
  - url: http://localhost:8080/is_coursework/api
    description: Local application server
...
paths:
...
  /journal/make-record:
    post:
      tags: [Journal]
      summary: Make record in journal for table
      description: Accessible to waiters, barmen, admins. Cooks cannot change table state.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalData'
      responses:
        '201':
          description: Journal record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalLogId'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /journal/get-table-status/{tableNumber}:
    get:
      tags: [Journal]
      summary: Get table status
      security:
        - BearerAuth: []
      parameters:
        - name: tableNumber
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TableNumber'
      responses:
        '200':
          description: Current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableStatus'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /journal/get-employee/{tableNumber}:
    get:
      tags: [Journal]
      summary: Get the waiter assigned to a table
      security:
        - BearerAuth: []
      parameters:
        - name: tableNumber
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TableNumber'
      responses:
        '200':
          description: Employee currently responsible for the table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /journal/get-all-statuses:
    get:
      tags: [Journal]
      summary: Get all table statuses
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Table status map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableStatusMap'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /journal/reset-employee:
    post:
      tags: [Journal]
      summary: Reset employee for a table
      description: Only administrators may reset employee.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalData'
      responses:
        '200':
          description: Assignment updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /journal/get-last-for-hours/{hours}:
    get:
      tags: [Journal]
      summary: Get journal records for last N hours
      security:
        - BearerAuth: []
      parameters:
        - name: hours
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JournalLogResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /order/get-orders-by-table/{tableNumber}:
    get:
      tags: [Orders]
      summary: Get orders for table
      security:
        - BearerAuth: []
      parameters:
        - name: tableNumber
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/TableNumber'
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
  /order/create:
    post:
      tags: [Orders]
      summary: Create an order for a table
      description: >-
        Waiters may create orders only for tables assigned to them. Administrators
        can create orders for any table.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreationData'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreationResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /order/{id}:
    get:
      tags: [Orders]
      summary: Get order by id
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /order/modify/{id}:
    put:
      tags: [Orders]
      summary: Modify order
      description: Only administrators may update orders.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /order/delete/{id}:
    delete:
      tags: [Orders]
      summary: Delete an order
      description: Only administrators may delete orders.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order deleted
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /order/get-all:
    get:
      tags: [Orders]
      summary: Get all orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '500': { $ref: '#/components/responses/ServerError' }
  /order/change-status/{id}:
    post:
      tags: [Orders]
      summary: Change order status
      description: >-
        Waiters, barmen, cooks, and admins can transition orders according to
        their role. Provide the new status via the query parameter.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Status updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
      additionalProperties: false
...
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        positions:
          $ref: '#/components/schemas/Positions'
        name:
          type: string
        secondName:
          type: string
        patronymic:
          type: string
          nullable: true
      required: [id, positions, name, secondName]
...
    JournalData:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        tableNumber:
          $ref: '#/components/schemas/TableNumber'
        tableStatus:
          $ref: '#/components/schemas/TableStatus'
      required: [employeeId, tableNumber, tableStatus]
    JournalLogId:
      type: integer
      format: int64
      description: Numeric identifier of the created journal log
    JournalLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tableStatus:
          $ref: '#/components/schemas/TableStatus'
        employee:
          $ref: '#/components/schemas/EmployeeResponse'
        tableNumber:
          $ref: '#/components/schemas/TableNumber'
        time:
          type: string
          format: date-time
      required: [id, tableStatus, employee, tableNumber, time]
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        time:
          type: string
          format: date-time
        journalLog:
          $ref: '#/components/schemas/JournalLogResponse'
        guestNumber:
          type: integer
          format: int32
        dish:
          $ref: '#/components/schemas/Dish'
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
      required: [id, time, journalLog, dish, orderStatus]
    OrderCreationData:
      type: object
      properties:
        tableNumber:
          $ref: '#/components/schemas/TableNumber'
        dishId:
          type: integer
          format: int64
        guestNumber:
          type: integer
          format: int32
      required: [tableNumber, dishId]
    OrderCreationResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        message:
          type: string
      required: [orderId, message]
...
    TableStatusMap:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/TableStatus'
    Dish:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        primeCost:
          type: number
          format: double
          readOnly: true
        cost:
          type: number
          format: double
        preparingTime:
          type: integer
          format: int32
        isKitchen:
          type: boolean
          description: True for kitchen dishes, false for bar items
      required: [name, cost, preparingTime, isKitchen]
...
    TableNumber:
      type: string
      enum: [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10]
    TableStatus:
      type: string
      enum: [free, occupied, not_paid, paid]
    OrderStatus:
      type: string
      enum: [accepted, cooked, delivered]
    Positions:
      type: string
      enum: [waiter, cook, barman, admin]